blueprints/gke/hpc/bundle/app/images/video-downloader/requirements.txt
internetarchive==3.3.0
google-cloud-storage==2.8.0
